requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
geopandas>=0.14
shapely>=2.0
jupyter>=1.0
ipykernel>=6.0
scikit-learn>=1.3
requests>=2.31.0
googlemaps>=4.10.0
google-genai>=0.2.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
neo4j>=5.0
openai>=1.0
fastapi>=0.110
uvicorn>=0.29
